finalizacao da parte conta

+++ spring04/src/main/java/br/itau/spring04/repository/ContaRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.itau.spring04.model.Conta;
 
-public interface ContaRepo extends CrudRepository<Conta, Long>{
+public interface ContaRepo extends CrudRepository<Conta, Integer>{
     public List<Conta> findAll();
 }
